refactoring: make BaseEntity type is generics and other entities id are Long

+++ shoppingOnline/src/main/java/com/roninEngineerF02/shoppingOnline/exception/CartNotFoundException.java
@@ -1,7 +1,7 @@
 package com.roninEngineerF02.shoppingOnline.exception;
 
 public class CartNotFoundException extends RuntimeException {
-    public CartNotFoundException(Integer userId) {
+    public CartNotFoundException(Long userId) {
         super("Cart not found for user ID: " + userId);
     }
 }
